+++ R/APtimer.R
@@ -1,25 +1,24 @@
 #' AP Timer
 #'
 #' Basic timer.
-#' @param marker Optional object of class \code{'POSIXct'}
+#' @param marker Optional object of class \code{'POSIXct'}.
+#' @param hush logical. If set to TRUE, this will silence printing to console.
 #' @return If \code{marker=NULL}, then \code{APtimer} returns an object of class \code{'POSIXct'}.
 #' When this same object is used as an input later on, then \code{APtimer} will print the time elapsed
-#' since it was evaluated
+#' since it was evaluated.
 #' @export
 
-APtimer <- function(marker = NULL){
+APtimer <- function(marker = NULL, hush = FALSE){
 
   # If no time marker is given, then create a new one
   if(is.null(marker)){
 
-    cat("Started timer at:", format(Sys.time(), "%Y-%m-%d, %X"), "\n")
+    if(!hush) cat("Started timer at:", format(Sys.time(), "%Y-%m-%d, %X"), "\n")
 
     return(proc.time())
 
   # If a marker is given, then return the time elapsed since that marker
   }else{
-
-    cat("Finished at:", format(Sys.time(), "%Y-%m-%d, %X"), "\n")
 
     elapsed.seconds <- (proc.time() - marker)["elapsed"]
 
@@ -30,6 +29,11 @@ APtimer <- function(marker = NULL){
     if(elapsed.seconds >= 3600){
       elapsed.print <- paste(round(elapsed.seconds/3600, 2), "hours")}
 
-    cat("Time elapsed:", elapsed.print,"\n")
+    if(!hush){
+      cat("Finished at:", format(Sys.time(), "%Y-%m-%d, %X"), "\n")
+      cat("Time elapsed:", elapsed.print,"\n")
+    }else{
+      return(elapsed.print)
+    }
   }
 }