+++ NEWS.md
@@ -1,3 +1,12 @@
+# APfun 0.1.1
+
+Added following functions:
+
+APcontours - Generates elevation contours at 'tidy' height values
+APprecise - Prints a number with all decimal values (i.e. no rounding)
+APopen - Opens a folder in Windows Explorer
+APpolygonize - Converts a raster into polygons using a GDAL utility
+
 # APfun 0.0.0.9001
 
 * Initial release of APfun